Further cleanups and additions
Pull a bit more from the MESS wiki, clean up the phrasing and update
outdated bits.

Also remove the 'current as of' date. It's meaningless until the
documentation is all completed.

Set up groundwork for a basic list of commonly used free frontends
suitable to new users.

Set up groundwork for a basic list of the absolutely most common Q&A
issues. Most of this should reference the MAME wiki down the road...